Added configuration to get the buttons that we want in the editor.
For now, the header button just toggles between formatting with h1 and formatting with p (to 'toggle off'). This is a little clunky, as the user could select another block of text before clicking again. On the bright side, the user will always be able to get the result they want with a second click. I'm not sure how to get the selection context in order to determine whether h1 is already applied, so I don't know how else to do this. The only other option would be to have no toggle off feature, but that's annoying too.