Removed the ng-class for adding the `datepicker-with-clear` and the ng-class for `btn-date-clear`.
Switched the `show-date-picker` span to wrap the entire component. This allows me to apply `inline-block` to the entire dateTime input. Which was causing an issue of not spanning full width. It also removes the need to style the entire component different if it has a clear button. Now it works like other input-groups.