Finished implementing input behavior. The timing here is tricky. We want to:
Create the control and assign the component validator to it (constructor) We want this as soon as possible so the control can get to the ngFormControl of the input.

Setup the validators (onInit) Here we have to wait until the inputs are setup so we can get the required text.

Subscribe to status changes and check the initial error state (afterViewInit) This has to happen last so ngControl has a chance to run the initial validation check in case the starting state has an error.

Since all of the configuration for the component validator are now getting set after the constructor, we can now inject it directly. Each component should then specify it's own provider so they'll get a new instance.