We would also like to have checkboxes tie into the form. To facilitate this, I'm splitting the base input component into two separate components. One is a 'decorator' class that adds validation logic the checkbox doesn't need. (The value can only be true or false, so there's not much to validate)
Unfortunately, using an ngModel on the component doesn't seem to play nicely with having an internal control, so I'm removing the base valueAccessor. We'll just have to standardize all inputs to support two-way binding to the `value`, which is easy because we can build it into the base input class.