The simple control-controlGroup relationship we're using for forms doesn't appear to tie the form 'valid' state with the controls. Instead we'll use a simple helper method to check the validity of all of the controls for now.
Use an angular 2 version of the form service to get the first error message off the form and notify it on a validation failure. As before, we could consider changing this if we wanted to eventually aggregate the errors.

One option would be to have the form service return an array of errors, and have the form notify each of them.