Dropped ngFormControl.
We're now essentially doing all the work in the input and validation input base classes, so I don't think we need to use ngFormControl anymore. Dropping it because it's possible this will try to duplicate some of the work and have a slight performance impact. For example, I thing ngFormControl will do an initial validation pass. But we have to do that manually for select and other components that don't use ngFormControl.