Created an angular 2 template renderer.
We want to wrap the complexity of taking a template ref and rendering it against a specified model. The new template renderer functions similar to the old one, but takes a template ref instead of an html string. This is primarily useful for components that want to stamp out dynamic content in their view.