Pull jquery and angular from systemJS
We want to pull as many dependencies from SystemJS as we can instead of shimming. We thought we had to shim angular before because the import wasn't working properly, but we found that if we referenced index.js instead of angular.js it works because index.js is a commonjs file instead of using globals.