Update bootstrapping for tests to use systemJS
We want to use systemJS when running tests so that we can make sure that only one instance of Angular 1 and Angular 2 is loaded during imports from separate packages. We now have a test shim (karma-test-shim) that imports all of the tests using system js and a setup file that runs the downgrade on the utilities package.