Move baseURL property  to the index page
We don't want to set the baseURL on the configuration because the config is shared by multiple consumers and the base URL is different when running tests (karma uses /base as a base url). Instead, we will set it in the index page just before it is needed. This way each consumer of the configuration can specify the base url that they need.