Includes tests in tests bundle
When building the bundle for running tests, we want to include the tests AND the test setup in the bundle. This simplifies the path resolution for systemJS because now everything is in the bundle and nothing has the /base karma prefix except the initial systemJS config file. We don't need to provide anything to karma now except the bundle and a few other polyfills.