Test fixes
Emit doesn't trigger the stream by default, so we'll spy directly on emit instead of subscribing to it.

Should be calledTwice after the second toggle for buttonToggle and checkbox.

Select also needs to mock componentValidator and busy for ngAfterViewInit. Also assert that busy is triggered with the options observable.

User rating needs to provide actual implementations of the object and guid services for ngOnInit. Don't assert against the value change, just against setValue. (setValue will change the value, but in this case it's mocked out)