Fix custom search feature and clean up timing issues
Removed the hasSearchOption variable.  The first custom option was being added correctly, but usage of the hasSearchOption flag was preventing subsequent custom options from being added.

Changed the return type of the create property from void to any.  This matches the interface definition.

Removed $q.when wrapper on call to this.getItems(), as the getItems function is defined to return a promise.  Wrapping a promise in $q.when is an anti-pattern.

Refactored branch of loadItems function that handles the useClientSearching case.  When useClientSearching is enabled, the API should be called one time to load the entire list, then the search text is filtered on the client.  However, it was possible for a user who types quickly to trigger the API call more than once.  Fixed this by caching the promise returned from the getItems function.