Implemented simpleCard and simpleCardList in angular 2
For consistency with the card container, we'll use attributes for the header, content, and footer. In the card container, the consumer will have to use a template tag so the content can be stamped out properly.

For autosave behavior, a card contains a form, and automatically triggers the form to save whenever it tries to close.