Inherit from form instead of composing it
In order for inputs to be able to find the 'form' of the simple card, we need to actually make the simple card inherit from form and provide itself as an instance of a form component. This kind of makes sense semantically since essentially we're saying a card is a type of form, and that's why you don't explicitly specify a form in its content.