Show/hide UI message w/o API call
These changes help resolve an issue that was causing the message log to continue displaying an uploading graphic after a failed upload.

The call to dataService.saveMessage was resolving immediately due to a promise anti-pattern inside the Attachments Service.  This caused an API call to be made via getTopPage before the new message was actually saved.  It did, however, incidentally cause the message being added to show up on the screen because the Attachments Service was temporarily adding it to the list when getMessages was called (again via getTopPage).

In a separate commit, the Attachments Service was updated to correctly return the actual promise for the file upload.  Subsequently, getTopPage was no longer being executed right away, and it made more sense for the MessageLog Service to temporarily add the message to the visibleMessages itself without making that API call through the Attachments Service.  Then the MessageLog Service also needs to remove the new message on failure, thus the need to import the arrayUtility.