Created an angular 2 dateFilter component
For now I'm only supporting the ability to go back by days. We _really_ don't want to build our own spinner here if we can help it. The behavior is _much_ easier and more straightforward this way.

Also, renaming selectedDate1 and selectedDate2 to dateFrom and dateTo to communication some additional meaning, since we're setting the range from the first to the second.