Added form baseInputs
This simplifies simpleCard by allowing us to remove save from the definition altogether, since it's a form base option. In the card container, bind save into the container. Then we'll pass it to the card. However, in this case we want to bind to the card container item, not the form value. For now we'll do this by overriding the submit with a function that sets the form value to the item, and then calls super.submit()