Implemented card header column
Angular 2 has no method of generating a template from an html string, that I can tell, so instead the consumer will have to specify column generic templates in the content of the rlCardContainer. The columns can optionally be identified by a name, which is then used on the template to differentiate the column it is to be used for.