Implemented container column header
Angular 2 has no method of generating a template from an html string, that I can tell, so instead the consumer will have to specify column generic templates in the content of the rlCardContainer. The columns can optionally be identified by a name, which is then used on the template to differentiate the column it is to be used for.

Also, currently it looks as if the 'let' statement is required for the name to not break. I'll have to figure out if there's a way to leave this off if it's not needed.