Use ViewChildren to manage the cards
We'll use ViewChildren so we can let angular manage the cards that should be included. Angular will out of the box handle only including the cards that are on the current view, which is something we would've run into an issue with unless we implemented on onDestroy hook to remove the cards from the container.