Converted the old long click button to use the new timeout service
We're running into issues in RL2 with the action timeout getting messed with before we can $timeout cancel it. I'm guessing it's some sort of issue with ngUpgrade. I'm hoping the new timeout service won't have the same issue, since it's not tied into angular. However, we do have to run a scope.$apply, since we're not using angular's timeout service anymore.