Clone the data source and filter builders
Because of the way angular 2 providers work, every card container builder in a given context is getting the same child providers. We don't have a good way around that right now because we're in an ngUpgrade context, so we can't set new providers very easily. For now we'll address by cloning the child builders.