Rename to rlNestedFormGroups to differentiate from actual forms
As of right now, we may not be able to use this behavior for nesting rlForms. It's possible we could if we used @ContentChildren, but I'm not going to try for that right now. The main thing I'd like to do with this right now is implementing validation groups in angular 2.