Inject a parent form and append the form to it
I figured out this trick when I was looking at the angular forms source code. The key to injecting a component of the same type is to use @SkipSelf() to tell angular to ignore the current component. With this fix in place, forms can now be nested as much as desired and errors will roll up to any parent forms.