Added a submitAndWait
My recent changes to submit inadvertantly broke the submit button. I'd forgotton that the submit button relies on the result of submit to show its spinner. However, it still seems to make sense that a consumer that's programmatically submitting the form should expect to get a true/false result indicating if the action was successful. To accomodate this, I'll add a submitAndWait function that does the same thing, but returns the wait value instead. We'll move as much as possible into shared functions to reduce duplicate code. The only thing that's really duplicated now is the if check against the validation result.