Refactoring waitOn functionality.
This functionality is starting to get reused more than I'd like. To resolve this, I'm adding an async helper with a function that turns a waitOn value into an observable. All other uses can simple consume this and subscribe to the new stream. For a primitive, the helper uses a behavior subject to return a stream with the single value included.