Tie a specific identifier to an offClick event instead of stopping propagation
When a document event is received, we can check to see if this is an event for the same off click or not. If so, we ignore it. This prevents any unexpected side effects from stopping propagation. For example, when multiple offClick directives were on a page, clicking on one would fail to trigger another, because event propagation was stopped.