Typeahead accessibility - first pass
Implemented accessibility by making it possible to tab to the popout list items, and to focus them with the up and down arrows. Since this requires some ugliness, I'm abstracting the list items out to a separate component. These have the ability to tell the list to focus the next or previous item (down or up arrow), and have a tab index of 0 to allow focus. The main ugliness to work out here is whether or not a tab index is acceptable, and how to clean up the focus next and previous logic. (Possibly with an array next() helper)