Enable up/down arrows without focus
Instead of using a tab index to actually enable focus on the list items, we'll create a simulated focus using a index that's incremented and decremented by the up/down arrows. The list tracks this index and displays an items as focused if it matches the index. On enter or blur (tab away) the currently 'focused' item is selected, and the focus is cleared. The beauty of this approach is that the actual focus never leaves the typeahead input, so the user can continue typing after they've started scrolling through the options.