Implemented a trigger directive
Since the typeahead was now adding a lot of custom behavior just to interface with the popout list, I thought it would be nice to add a popout trigger directive to handle this. The downside is this necessitated adding some complexity by using a service to communicate between the popout components. However, I was able to get it working so the typeahead just has to drop in the trigger component and register the service as a provider and it works like magic