Added a tabindex
Since the popout list uses focus at its main open/close trigger now, we need to make select focusable. This is a bit of a hack, but unfortunately it seems to be the only way to make an element focusable for browsers. On the bright side, an index of 0 seems to cause it to fall into the natural tab order.