Blur the element on a select event
I was running into a minor issue on the select component where:
The user focuses the input, then mouses down to an item.
The user hits enter to select the item.
The items close, but focus is still on the input, forcing the user to click away then back again.
I noticed that the input does unfocus when the user clicks an item, so I thought it made sense to do the same for a keyboard select.

Verified that a 'tab' still correctly focuses the next element.