Click fixes
I ran into a really tricky bug where the blur event suppresses the click on the popout items, which causes the items to not be clickable at all.
Instead of using a click event on the popout item, use a mouseenter and mouseleave event to set or clear the current focus item. That way when a blur happens on click, the highlighted item will still be selected.