Tidy up ng1 buttonAsync promises
Promise rejections in the action trigger should be rethrown, otherwise they will be swallowed.  In the first iterartion of these changes, throwing the error caused issues in the unit tests, so it was skipped.  Changed the buttonAsync to return a promise from the trigger function.  Changed tests to use mock promises and assert against the rethrown error.  Also removed unnecessary calls to scope.$digest in unit tests that were not using promises.