Simplified data source testing using the angular test tooling
This is where the tooling really shines. We can simply drop in all of the utility providers and we gain the ability inject the data source processor without caring about some of the utility deps and the ordering of what dependencies have dependencies of their own.

Requires making the processor injectable for the sake of testing. This seems worthwhile for the relatively low cost.