Refactor synchronized requests logic for async data sources
rxjs natively supports ignoring all but the latest request. Instead of injecting a service to handle this, we'll just convert any promises to observables and synchronize via rxjs. This removes a dependency and also allows us to support observable requests in addition to promises.

Should have no client-facing impact. Unit tests prove that it works as expected.