Injection conflict shim
Ugly though this is, it seems that the injections for simple card, card, and selectable card are clashing. I'm guessing it has something to do with the fact that they all inherit from form. In any case, decorating a parameter with @Inject() seems to screw up the parameter in the same 'index' on the other classes, unless it shares the same type. For this reason I'm filling in the 5th parameter to simple cards with a null injection so card and selectable card can correctly inject the card container.