Created an input component for entering signatures
Uses jSignature to render a signature control on the screen. Unfortunately this has the same challenges as rlSpinner, where we have to re-render the control every time the canvas div element becomes visible again. The nice thing is that this one seems to nicely handle sizing to the parent container, which the other one didn't do very well.