Casting setTimeout() to any.
Currently in the core project we are using the AWS S3 SDK and it has a dependency on the @types/node which when we go to compile throws an error that we cannot cast this number to type Timer. I would prefer this to be a temporary solution until something better is found.