Created a button route component
This component works like a button link except that it uses a routerLink instead of href, so it has better integration with the angular router. Also, it supports the routerLinkActive through the activeClass property. (Defaults to 'active', it doesn't appear to be possible to conditionally enable using the active class)