Created an initial version of the new builder
The new builder will act as a factory for card container 'instances' instead of an instance itself. This solves the issue where you couldn't include two card containers on the same page without cloning the builder. The 'getInstance' function will take most of the simple settings, leaving only the data source, filters, and columns to be specified individually.