Created a new data source for representing the data as an observable stream
It seems like this is really the only 'simple' data source we'll ever need. This can replace the old data service data source, as well as the simple array data source. Also, the consumer gets finer-grained control because they could pass in any data stream and manipulate it how they want. The only thing this doesn't accomplish is any server integration, which we'll still do through the smart data source.