Reimplemented selectable card containers, minus the selection control
This is tricky because we still want this to be fully reactive, which means we need to be able to map the data set to a 'selection state' data set. To accomplish this, we'll wrap the data set items with a selection object. This requires having some way to identify the data set items, since the instances will change. We'll require the items to include an 'id' property, at least for the time being.