Add internal Validators to validationInput
Internal validators are validators that can only be set by the input's controller, making it more convenient for the controller to specify its own validators.

It was already possible for input controllers to specify their own validators, but to do so would require some fairly complicated logic to check if the validators property is defined before attempting to append the custom validators.