Added paging data to the final event model
This allows this to translate directly to the params for a server search request. Eventually, we could wire in a paging stream to configure it so a request to get an additional set of pages is fired once we get to a certain page point. For now, it works just as the other one did, where it always gets the first 200 items.

This allows us to type the observable streams now.