Differentiate initial vs 'active' stream
For the initial, we'll just fire one request immediately, and suppress all additional requests. For the 'active' stream. We'll skip 1 event for every throttle change. This is because the throttle change is always going to be triggered by the resolution of a request, so we don't want to immediately kick off another one.