Implemented a new smart data source
Uses the smart data actions to load data when the filters change. On init, the source fires an initial request (toRequestStream with initial set to true) to 'bootstrap' the stream. Then When that request completes, we initiate a 'normal' execution stream, so the data can refresh whenever the necessary filters, sorts, and throttle results change.