Keep the spinner if the data set is null
Generally, the data set should at least have an empty array. If it has null instead, we'll keep the spinner showing. This works as the observable data source does. So if the consumer wants to externally trigger the loading to restart, they can push an event to the stream with a null data set and the spinner will show up again.