For a one-time subscription, use `.first()` instead of subscribing and unsubscribing
This is much simpler and better than managing the subscription ourselves. Unfortunately, `.first()` doesn't help in cases where the subscription directly affects the stream. (Like in the pager's 'increment/decrement' this stream by on). In this case, I'll add a `.delay(0)` for the moment to force it to wait one tick so the `.first()` can complete the stream. We'll review this in the mob pairing session.