Drop the distinctUntilChanged()
This doesn't actually seem to suppress duplicate events at the moment. I'm guessing it's because it uses '===' comparison and the new event coming through is a different object, even if it looks the same. At any rate, duplicates should get caught by the debounce if they are fired at the same time. If the user reselects the same filter value, it might trigger the source to reload, which isn't ideal, but that's something we can look into later.