Use a subject for showContent
Change detection is broken for closing cards now that on push detection is enabled. This is because when a card opens, other cards that are closing have no way of knowing they need to update. By switching showContent to an observable, the async pipe will take care of this for us.