Created a component for defining a simple card container
One of the requests for the team was to make it easier to build a simple card container with less setup that has to go into it. We can do this by having a component that inputs the data, columns, and optional options, and sets up a card container builder for you. To allow for live data changes, we'll take the data input and convert it to a behavior subject, and provide this to an observable data source for the card container.

The biggest hurdle with this approach is that templates are kind of finicky. To get around this, we have to @ContentChild(ren) the templates into the simple container, then bind these into the card container as external templates. The card container then has to check to see if it needs to use external templates instead of the internal ones.