Updated typeahead to use waitOn
The typeahead spinner should wait on the load request and then subscribe to it to open the list and show the visible items. With the direction we're moving, it seems like a `.subscribe()` should always be the end of a call chain, so we don't want to return the observable here.

For the tests, we can just subscribe to visibleItems$ to get the current state of the visible items.