Clean up form subscriptions
The form shouldn't subscribe to mark the form as pristine. Instead we'll apply this as a `.do()` on the save request.

Also, for the `.submit()` function, since this is the end of the call chain, we'll kick off the subscription here. This will handle the subscription 'kickoff' for cards and simple cards.