submitAndClose shouldn't subscribe to the wait value
This is a non-action point, which is just an external API someone else can use to trigger a save followed by a close of the dialog. For this case, the consumer that triggers this will have to actually make the subscription. For most use cases, this would just be the 'save and close' async button.