Added the ability to focus textboxes and typeaheads
We'll likely need to able to focus most or all inputs, but we'll just start with these for now. This gives us the ability to call into these input components to programatically focus the input, which is useful for when an action leads to a particular value needing to be edited. Jeff and I discussed this and various options, and decided to go with this as the most straightforward approach.