Popout list should only hide the items if both the custom and list items are empty
This resolves an issue where when the list items are empty, the custom items are hidden because isEmpty returns true. An example is when using a typeahead if the search returns no results. If creating custom items is enabled, the search value should always show up at the top as an available option, whether or not any results are returned.