Check against the data options, not the list items
This was causing a circular dependency issue. The list items wouldn't renter until isEmpty is false, but isEmpty will never become false until list items are present. We can still check against the custom items because they'll always be present as content children, whether the list is included or not.