Updated the button components to work with the new upgrade adapter
Anything using the old upgradeAdapter doesn't seem to work anymore. Instead we create a directive that extends UpgradeComponent. This does the upgrade for us.

Converted the ng2 buttons bootstrapper to an actual ng2 component. Otherwise we'd have to figure out how to downgrade the angular 2 components, which I wasn't ready to tackle yet.