Shim the angular 1 reference from es6 to the global scope instead of the other way around
The other way was causing issues with conflicting versions of jquery. When components need to import a jquery extension and have it decorate the jquery object, it wasn't decorating the correct jquery reference. By doing this, we can have it running the way it was, but still make the upgrade module happy by making angular 1 available as a global