Fixed the test setup so we can run tests using the new upgrade module
Like before, we still need to bootstrap ng-upgrade and save a reference to the injector so we can get at it from angular 1 components and services. In this case, we bootstrap it via ng2 using the browser dynamic testing platform. Since the downgrade is static, we no longer need to call the downgrade utilities and components function.