Added support for setting an offClick directive as inactive
The directive will watch for ngOnChanges and add and remove the event listener as needed. This is important because we might be going back to using off click for closing the popout list in our select and typeahead dropdowns. Without this change, that would mean that every dropdown and typeahead adds a document event to the page. Instead we can bind to offClickActive to only turn on the event listener when the dropdown is open, which will restrict us to one document listener at a time. (Only one dropdown should be able to open at a time)