Don't use focus and blur as the open/close triggers for the list
This was causing issues with the list closing when the user tried to click on the scroll bar for the list. Instead we'll go back to using an off click trigger to close the list. The list will also close on selection. Also, the user can now toggle the list by clicking on the input itself, so they can open and close it as needed by that method. This should actually fix the gimicky behavior we were occasionally seeing in the importer, when somehow focus was lost but the list was still visible. The input no longer needs to be focused in order to select an item from the list, as long as the list is open.

Slight feature regression: tab will no longer select the currently highlighted item in the list. I tried to add a keyup.tab event listener, but it didn't seem to work. If we want to get this behavior back, we'll have to look at how angular does tab events.