Created a stateless message log component
This is purely a view component that shows the log portion of the message log. The consumer will pass in the list of messages to be shown, along with information regarding what page the log is on. The log component will emit paging events when the 'older', 'top', or 'refresh' buttons are clicked. These events will be handled somewhere above for sending requests to the server to get the log messages.