Updating typeahead so it always returns it's results.
in the Test Equipment Typeahead issues if you are to type RS get the results the add L to your search it hangs indefintitly because the list is not calling the open/detectChanges code. Because the Result set is the same as the previous query. by moving these actions out side of the coniditional it should cause results to always be displayed.