Fix for opening a dialog from a dialog
When one dialog wants to open from another, currently they both get closed immediately, since the second dialog opens and then the first one immediately closes. I tried adding a timeout or changing the order so the old dialog closes first, but this only seems to result in two overlays. Instead, we can actually fix this by disabling close events from a dialog unless it is the currently active dialog. If a second dialog opens up, the first dialog gets marked as inactive and close events are ignored from it. This way, the dialog outlet template simply gets replaced and the dialog outlet doesn't need to be closed and reopened.