Clear the active dialog context on close
Currently, if the dialog is closed, the isOpen() method on the dialog component will return true as long as it was the last opened dialog. The easiest way to fix this is to wipe out the context when the dialog closes, so we can always tell that if there is a dialog context set, that dialog is open.