Type of button
On certain pages, short clicking a long click button is causing the form to get submitted. This is because by default buttons are tied to the submit action unless the button type is explicitly set to 'button'. This usually isn't an issue for us because it only affects where we use ASP.NET forms with a default post action.

https://renovo.myjetbrains.com/youtrack/issue/RLv2-2049